+++ public/service-worker.js
@@ -1,6 +1,6 @@
 // /public/service-worker.js
 
-const CACHE_NAME = 'outerpedia-cache-v1.27.3';
+const CACHE_NAME = 'outerpedia-cache-v1.28.0';
 const urlsToCache = [
   '/',
   '/manifest.json',